put delete duplicates feature here
bc i think the best option for this would be with the menu, and this isn't added in the delete duplicates/main branch. i think i could start a branch from here, but then the delete duplicates would have more unfinished not properly working code. although i could then merge this. idk. maybe not the best decision but it doesn't have impact really