Shuffle Reset Setting
update for the main application lol
I added a setting to put the current song at the top when shuffling the playlist.
Before the current song was always put at the top. I made it like that cause my music app does this even tho  don't like it that way. That's why I made that setting. I could also do a feature request for this music app I use... hmm 🤔